Style: 웹캠 페이지 타입스크립트화
