+++ src/App.tsx
@@ -3,11 +3,11 @@ import './App.css';
 import WebCamPage from './webCam';
 
 function App() {
-  return (
-    <div>
-        <WebCamPage />
-    </div>
-  );
+    return (
+        <div>
+            <WebCamPage />
+        </div>
+    );
 }
 
 export default App;
