+++ src/common/constraints.ts
@@ -1,0 +1,7 @@
+export default {
+    video: {
+        width: 800,
+        height: 600,
+    },
+    audio: false,
+};
